requirements.txt
python-telegram-bot==21.3
openai-whisper
groq
python-dotenv
